requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.12.3
attrs==25.3.0
Authlib==1.6.1
Babel==2.9.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==25.1.0
blis==0.7.11
Brotli==1.1.0
cached-property==2.0.1
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorlog==6.9.0
confection==0.1.5
cryptography==3.4.8
cyclopts==3.22.3
cymem==2.0.8
decorator==4.4.2
distlib==0.4.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.17
ebaysdk==2.1.5
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.10.6
ffmpy==0.6.1
filelock==3.18.0
freezegun==1.1.0
fsspec==2025.7.0
geoip2==2.9.0
gevent==21.8.0
google==3.0.0
google-ads==25.0.0
google-ai-generativelanguage==0.6.15
google-analytics-admin==0.24.1
google-analytics-data==0.18.19
-e git+https://github.com/googleanalytics/google-analytics-mcp.git@3f849bbc24f40fc86aa83e470bfa1ce2d9631616#egg=google_analytics_mcp
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.8.5
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gradio==5.41.1
gradio_client==1.11.0
greenlet==1.1.2
groovy==0.1.2
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
hubspot-api-client==9.0.0
huggingface-hub==0.34.3
idna==2.10
isodate==0.7.2
Jinja2==3.0.3
joblib==1.4.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
libsass==0.20.1
lxml==4.8.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
maxminddb==2.7.0
mcp==1.12.2
-e git+https://github.com/GunnarGriese/mcp-server-google-analytics.git@1e9f3ae637dc8815191e188d973dd3816df9cf05#egg=mcp_server_google_analytics
mdurl==0.1.2
murmurhash==1.0.10
mypy_extensions==1.1.0
nltk==3.8.1
nox==2022.1.7
num2words==0.5.10
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ofxparse==0.21
openapi-pydantic==0.5.1
orjson==3.11.1
packaging==24.1
pandas==2.3.1
paramiko==3.4.0
passlib==1.7.4
pathspec==0.12.1
pbkdf2==1.3
Pillow==9.0.1
pipx==1.7.1
platformdirs==4.3.8
polib==1.1.1
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.0
psycopg2==2.9.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==1.4.2
pydub==0.25.1
Pygments==2.18.0
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.1.2
PyPDF2==1.26.0
pyperclip==1.9.0
pyserial==3.5
pysftp==0.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ldap==3.4.0
python-multipart==0.0.20
python-stdnum==1.17
pytz==2025.2
pyusb==1.2.1
PyYAML==6.0.2
qrcode==7.3.1
referencing==0.36.2
regex==2024.5.15
reportlab==3.6.8
requests==2.25.1
requests-file==2.1.0
requests-oauthlib==1.2.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-rst==1.3.1
rjsmin==1.1.0
rpds-py==0.26.0
rsa==4.9
ruff==0.12.7
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sse-starlette==3.0.2
starlette==0.47.2
thinc==8.2.5
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.66.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.5
userpath==1.9.2
uvicorn==0.35.0
virtualenv==20.32.0
vobject==0.9.6.1
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==2.0.2
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
zope.event==5.1
zope.interface==7.2
